+++ packages/core/services/index.ts
@@ -13,14 +13,10 @@ export { default as DatasetService } from "./DataSourceService";
 export type { default as ExecutionEnvService } from "./ExecutionEnvService";
 export { ExecutableEnvCancellationToken, SystemDefaultAppLocation } from "./ExecutionEnvService";
 export type { SaveLocationResolution } from "./ExecutionEnvService";
-export type {
-    default as FileDownloadService,
-    DownloadResult,
-    FileInfo,
-} from "./FileDownloadService";
+export type { DownloadResult, FileInfo } from "./FileDownloadService";
 export { DownloadResolution, FileDownloadCancellationToken } from "./FileDownloadService";
 export { default as FileService } from "./FileService";
-export { default as FileStorageServiceBase } from "./FileStorageServiceBase";
+export { default as FileDownloadService } from "./FileDownloadService";
 export type { default as FileViewerService } from "./FileViewerService";
 export { FileViewerCancellationToken } from "./FileViewerService";
 export { default as HttpServiceBase } from "./HttpServiceBase";
